Fix oapv_bsw_write1() to add necessary typecast (#189)
13d630800b316dfd84ad2c1503cb41a26aa258b5 changed the bitstream parse and write functions
to use a 64bit variable in the bitstream structure.

In that CL, oapv_bsw_write() was updated to typecast the value being written before
shifting but, but oapv_bsw_write1() wasn't updated.

This CL updates oapv_bsw_write1() to add the necessary typecast to u64.

+++ src/oapv_bs.c
@@ -100,7 +100,7 @@ int oapv_bsw_write1(oapv_bs_t *bs, int val)
     oapv_assert(bs);
 
     bs->leftbits--;
-    bs->code |= ((val & 0x1) << bs->leftbits);
+    bs->code |= ((u64)(val & 0x1) << bs->leftbits);
 
     if(bs->leftbits == 0) {
         oapv_assert_rv(bs->cur < bs->end, -1);
